WIP api docs

+++ update/sets.mjs
@@ -36,9 +36,11 @@ const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const DATA = path.resolve(__dirname, '../data');
 
 for (const file of fs.readdirSync(path.join(DATA, 'analyses'))) {
+  if (file === 'index.md') continue;
   fs.unlinkSync(path.join(DATA, 'analyses', file));
 }
 for (const file of fs.readdirSync(path.join(DATA, 'sets'))) {
+  if (file === 'index.md') continue;
   fs.unlinkSync(path.join(DATA, 'sets', file));
 }
